refactor: replaces old solana module with current module
