--- coins/karl_solana/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "karl_solana"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-name = "karl_solana"
-crate-type = ["cdylib", "lib"]
-path = "src/lib.rs"
-
-#Solana-specific because main shouldn't be called by this module
-[features]
-no-entrypoint = []
-
-[dependencies]
-async-trait = "0.1.60"
-
-solana-client = "1.16.8"
-solana-sdk = "1.16.8"
-walletd_coin_core = { path = "../core/", version = "0.2" }
-walletd_mnemonics_core = { path = "../../mnemonics/core", version = "0.2" }
-thiserror = "1.0.47"
-tokio = "1.14.1"
-
-[workspace]
