+++ coins/solana/examples/accounts.rs
