+++ coins/solana/examples/get_rent_parameters.rs
