+++ coins/solana/examples/keypairs.rs
