+++ coins/solana/src/error.rs
