+++ coins/solana/src/lib.rs
@@ -1,70 +1,40 @@
-use core::fmt;
-use core::fmt::Display;
 
-use base58::ToBase58;
-use solana_client::rpc_client::RpcClient;
-// const URL: &str = "https://api.devnet.solana.com";
-use walletd_hd_key::HDNetworkType;
+//#![deny(missing_docs)]
+#![cfg_attr(not(test), forbid(unsafe_code))]
+//pub use crate::solanaclient as SolanaClient;
+pub mod solanaclient;
+use solana_sdk::entrypoint::ProgramResult;
+//use solana_sdk::bpf_loader::id as bpf_loader_id;
 
-#[derive(Default)]
-pub enum SolanaFormat {
-    #[default]
-    Standard,
-}
+/// An ERC20-like Token program for the Solana blockchain
 
-impl Display for SolanaFormat {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            SolanaFormat::Standard => write!(f, "Standard"),
-        }
-    }
-}
+mod error;
+pub use error::Error;
+// pub mod instruction;
+// pub mod native_mint;
+// pub mod processor;
+// pub mod state;
 
-// Solana uses accounts and transactions. Hard to map the concept to a "Wallet" per say
-// TODO: Eventually provision functionality to use walletd_keystore to restore from a keypair file
-// * a base58-encoded public key
-// * a path to a keypair file
-// * a hyphen; signals a JSON-encoded keypair on stdin
-// * the 'ASK' keyword; to recover a keypair via its seed phrase
-// * a hardware wallet keypair URL (i.e. usb://ledger)
 
-pub struct SolanaWallet {
-    address_format: SolanaFormat,
-    public_address: String,
-    private_key: String,
-    public_key: String,
-    network: HDNetworkType,
-}
+use solana_sdk::pubkey::Pubkey;
+use solana_sdk::program_error::ProgramError;
+use solana_sdk::declare_id as solana_declare_id;
 
-impl SolanaWallet {
-    pub fn public_address_from_public_key(public_key: &Vec<u8>) -> String {
-        public_key.to_base58()
-    }
+#[cfg(not(feature = "no-entrypoint"))]
+mod entrypoint;
 
-    pub fn keypair_base58(private_key: &[u8; 32], public_key: &[u8; 33]) -> String {
-        let mut keypair = [0u8; 64];
-        keypair[0..32].copy_from_slice(&private_key.as_slice()[0..32]);
-        keypair[32..64].copy_from_slice(&public_key.as_slice()[1..33]);
-        keypair.to_base58()
-    }
-}
+// Export current sdk types for downstream users building with a different sdk version
+// pub use solana_program;
+// pub mod solana_sdk;
 
-impl Display for SolanaWallet {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        writeln!(f, "Solana Wallet")?;
-        writeln!(f, " Network: {}", self.network)?;
-        writeln!(f, " Private Key: {}", self.private_key)?;
-        writeln!(f, " Public Key: {}", self.public_key)?;
-        writeln!(f, " Address Format: {}", self.address_format)?;
-        writeln!(f, " Public Address: {}", self.public_address)?;
-        Ok(())
-    }
-}
 
-pub struct BlockchainClient(pub RpcClient);
+// TODO: move this for other modules, but use client and sdk instead of program
+// use solana_program::{entrypoint::ProgramResult, program_error::ProgramError, pubkey::Pubkey};
 
-impl BlockchainClient {
-    pub fn new(url: &str) -> Result<Self, anyhow::Error> {
-        Ok(Self(RpcClient::new(url)))
-    }
-}
+// /// Checks that the supplied program ID is the correct one for SPL-token
+// pub fn check_program_account(spl_token_program_id: &Pubkey) -> ProgramResult {
+//     if spl_token_program_id != &solana_sdk::bpf_loader::id() {
+//         return Err(ProgramError::IncorrectProgramId);
+//     }
+//     Ok(())
+// }
