+++ coins/solana_old/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "walletd_solana"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+authors = ["The WalletD Team"]
+description = "Solana walletd library"
+homepage = "https://walletd.org/"
+repository = "https://github.com/walletd/walletd"
+readme = "README.md"
+categories = ["cryptography::cryptocurrencies"]
+
+[lib]
+name = "walletd_solana"
+path = "src/lib.rs"
+
+[dependencies]
+
+walletd_bip39 = { path = "../../mnemonics/bip39", version = "0.1" }
+walletd_hd_key = { path = "../../key_manager/hd_key", version = "0.1" }
+
+anyhow = "1.0.68"
+base58 = { version = "0.2.0" }
+ed25519-dalek-bip32 = "0.2.0"
+hex = { version = "0.4.2" }
+solana-client = "1.14.7"
+solana-sdk = "1.14.10"
